requirements.txt
flask
flask-restful
flask-sqlalchemy
pymysql
flask-migrate
flask-script
flask-jwt-extended
requests
pytest
flask-cors
pytest-cov
coverage
mock
passlib
python-dotenv